Several updates
I use this plugin heavily and was looking for a feature to store every person into it's dedicated directory rather than just a note.
For that I implemented a dropdown that lets you choose if you want to store notes for every person the "default" way (one note per person), by lastname, or one directory per person (in which you can add more notes). Additionally, I introduced a bugfix to correctly load settings and represent them in the ui and the possibility to make multi line descriptions using document fragments.